refactor http types into their own files
