+++ cpm-hub/http/Endpoint.h
@@ -24,6 +24,7 @@
 
 #include <infrastructure/Optional.h>
 #include <http/http.h>
+#include "HttpParameterMap.h"
 
 
 class Endpoint {
