+++ cpm-hub/http/HttpRequest.h
@@ -18,19 +18,15 @@
 #pragma once
 
 #include <string>
-#include <infrastructure/Filesystem.h>
-#include <authentication/Authenticator.h>
 
 
-class BasicAuthenticator: public Authenticator {
-public:
-    BasicAuthenticator(Filesystem *filesystem);
+struct HttpRequest {
+    std::string method;
+    std::string body;
+    struct HttpParameterMap parameters;
+    struct HttpParameterMap headers;
 
-    void setAccessFile(std::string filename);
-
-    bool authenticate(const char *key);
-
-private:
-    Filesystem *filesystem;
-    std::string access_file;
+    HttpRequest(std::string _body="") {
+        body = _body;
+    }
 };
